ujednolicenie do jednego UserDbService
