+++ src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -3,7 +3,7 @@ package com.kodilla.ecommercee.controller;
 import com.kodilla.ecommercee.domain.dto.UserDto;
 import com.kodilla.ecommercee.exceptions.UserExistsException;
 import com.kodilla.ecommercee.exceptions.UserNotFoundException;
-import com.kodilla.ecommercee.service.UserService;
+import com.kodilla.ecommercee.service.UserDbService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -16,19 +16,19 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final UserDbService userDbService;
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UserDto> addUser(@RequestBody UserDto userDto) throws UserExistsException {
-        return ResponseEntity.ok(userService.createUser(userDto));
+        return ResponseEntity.ok(userDbService.createUser(userDto));
     }
     @PutMapping(value = "/changeStatus/{userId}")
     public ResponseEntity<UserDto> changeUserStatus(@PathVariable long userId) throws UserNotFoundException {
-        return ResponseEntity.ok(userService.changeUserStatus(userId));
+        return ResponseEntity.ok(userDbService.changeUserStatus(userId));
     }
 
     @PutMapping(value = "/createKey/{userId}")
     public ResponseEntity<UUID> createUserKey(@PathVariable long userId) throws UserNotFoundException {
-        return ResponseEntity.ok(userService.createUserKey(userId).getUserKey());
+        return ResponseEntity.ok(userDbService.createUserKey(userId).getUserKey());
     }
 }
